% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_files.R
\name{convert_to_ods}
\alias{convert_to_ods}
\title{Save a copy of an xlsx file as an ods file}
\usage{
convert_to_ods(path)
}
\arguments{
\item{path}{path to xlsx file; can be either a relative or absolute file path}
}
\description{
Uses VBA code to save a copy of an xlsx file to the accessible ODS format, retaining all sheets and formatting.
}
\details{
Files converted can be via a relative (from working directory) or absolute (full) file path. In either case, the output ODS file will be returned in the same folder as the XLSX file.
}
