% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vbs_execute}
\alias{vbs_execute}
\title{Execute a VBS script including arguments}
\usage{
vbs_execute(vbs_file, ...)
}
\arguments{
\item{vbs_file}{Name of a vbs script file saved as part of the package}

\item{...}{Arguments to be passed to the specified VBS script}
}
\description{
Create a filepath for a referenced VBS file
}
\details{
Formats the name of a VBS file to be used
}
