% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vbs_file_path}
\alias{vbs_file_path}
\title{Formats the name of a VBS script file into the filepath within the package}
\usage{
vbs_file_path(vbs_file)
}
\arguments{
\item{vbs_file}{Name of a vbs script file saved as part of the package}
}
\description{
Create a filepath for a referenced VBS file
}
\details{
Formats the name of a VBS file to be used
}
